chore: moved expression builder logic at function builder to its own builder
